+++ src/lib/components/model-picker/index.ts
@@ -1,0 +1,3 @@
+import ModelPicker from './model-picker.svelte';
+
+export { ModelPicker };
